cleanup: add dedicated singleton header
Use a single, standard singleton template instead of copy and pasting
the code for each use.
