+++ src/modules/packetio/stack/dpdk/tcpip_mbox.hpp
@@ -4,26 +4,11 @@
 #include <memory>
 
 #include "packetio/stack/dpdk/sys_mailbox.hpp"
+#include "utils/singleton.hpp"
 
 namespace openperf::packetio::dpdk {
 
-template <typename T> class singleton
-{
-public:
-    static T& instance()
-    {
-        static T instance;
-        return instance;
-    }
-
-    singleton(const singleton&) = delete;
-    singleton& operator=(const singleton) = delete;
-
-protected:
-    singleton(){};
-};
-
-class tcpip_mbox : public singleton<tcpip_mbox>
+class tcpip_mbox : public utils::singleton<tcpip_mbox>
 {
 public:
     sys_mbox_t init();
